requirements.txt
selenium 
beautifulsoup4 
requests
anthropic
python-dotenv
undetected-chromedriver
openai